move STATUSES and DEFAULT_CONFIG into scriptworker.constants
Also, change `DEFAULT_CONFIG` into a frozendict and rename
`list_to_tuple` to `freeze_values`; dict values will be changed
into frozendicts.
