+++ scriptworker/client.py
@@ -1,10 +1,5 @@
 #!/usr/bin/env python
 """Scripts running in scriptworker will use functions in this file.
-
-Attributes:
-    STATUSES (dict): maps taskcluster status (string) to exit code (int).
-    REVERSED_STATUSES (dict): the same as STATUSES, except it maps the exit code
-        (int) to the taskcluster status (string).
 """
 import json
 import jsonschema
@@ -12,20 +7,8 @@ import os
 import re
 from urllib.parse import urlparse, unquote
 
-from scriptworker.config import DEFAULT_CONFIG
+from scriptworker.constants import DEFAULT_CONFIG, STATUSES
 from scriptworker.exceptions import ScriptWorkerTaskException
-
-
-STATUSES = {
-    'success': 0,
-    'failure': 1,
-    'worker-shutdown': 2,
-    'malformed-payload': 3,
-    'resource-unavailable': 4,
-    'internal-error': 5,
-    'superseded': 6,
-}
-REVERSED_STATUSES = {v: k for k, v in STATUSES.items()}
 
 
 def worst_level(level1, level2):
