+++ scriptworker/test/test_integration.py
@@ -11,7 +11,8 @@ import os
 import pytest
 import slugid
 import tempfile
-from scriptworker.config import CREDS_FILES, DEFAULT_CONFIG, read_worker_creds
+from scriptworker.config import CREDS_FILES, read_worker_creds
+from scriptworker.constants import DEFAULT_CONFIG
 from scriptworker.context import Context
 import scriptworker.log as swlog
 import scriptworker.worker as worker
@@ -49,7 +50,7 @@ To skip integration tests, set the environment variable NO_TESTS_OVER_WIRE""".fo
 
 def build_config(override, basedir):
     randstring = slugid.nice()[0:6].decode('utf-8')
-    config = deepcopy(DEFAULT_CONFIG)
+    config = dict(deepcopy(DEFAULT_CONFIG))
     GPG_HOME = os.path.join(os.path.basename(__file__), "data", "gpg")
     config.update({
         'log_dir': os.path.join(basedir, "log"),
